feat: initial prototype of unification
