+++ client/src/components/Home/HomeGridUQScore/HomeGridUQScore.tsx
@@ -1,8 +1,7 @@
 import { watchAccount } from '@wagmi/core';
-import uniq from 'lodash/uniq';
 import React, { FC, useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useAccount, useConnectors } from 'wagmi';
+import { useAccount } from 'wagmi';
 
 import { wagmiConfig } from 'api/clients/client-wagmi';
 import HomeGridUQScoreAddresses from 'components/Home/HomeGridUQScore/HomeGridUQScoreAddresses';
@@ -16,13 +15,10 @@ import {
   GITCOIN_PASSPORT_CUSTOM_OCTANT_DASHBOARD,
   TIME_OUT_LIST_DISPUTE_FORM,
 } from 'constants/urls';
-import useCheckDelegation from 'hooks/mutations/useCheckDelegation';
 import useRefreshAntisybilStatus from 'hooks/mutations/useRefreshAntisybilStatus';
 import useAntisybilStatusScore from 'hooks/queries/useAntisybilStatusScore';
 import useCurrentEpoch from 'hooks/queries/useCurrentEpoch';
 import useUqScore from 'hooks/queries/useUqScore';
-import useUserTOS from 'hooks/queries/useUserTOS';
-import toastService from 'services/toastService';
 import useDelegationStore from 'store/delegation/store';
 
 import styles from './HomeGridUQScore.module.scss';
@@ -33,9 +29,7 @@ const HomeGridUQScore: FC<HomeGridUQScoreProps> = ({ className }) => {
     keyPrefix: 'components.home.homeGridUQScore',
   });
   const { address, isConnected } = useAccount();
-  const connectors = useConnectors();
   const { data: currentEpoch } = useCurrentEpoch();
-  const { data: isUserTOSAccepted } = useUserTOS();
 
   const { data: uqScore, isFetching: isFetchingUqScore } = useUqScore(currentEpoch!);
 
@@ -44,7 +38,6 @@ const HomeGridUQScore: FC<HomeGridUQScoreProps> = ({ className }) => {
     useState(false);
 
   const {
-    delegationPrimaryAddress,
     delegationSecondaryAddress,
     primaryAddressScore,
     isDelegationInProgress,
@@ -57,7 +50,6 @@ const HomeGridUQScore: FC<HomeGridUQScoreProps> = ({ className }) => {
     setDelegationPrimaryAddress,
     setDelegationSecondaryAddress,
     setIsDelegationCalculatingUQScoreModalOpen,
-    setIsDelegationCompleted,
   } = useDelegationStore(state => ({
     delegationPrimaryAddress: state.data.delegationPrimaryAddress,
     delegationSecondaryAddress: state.data.delegationSecondaryAddress,
@@ -68,7 +60,6 @@ const HomeGridUQScore: FC<HomeGridUQScoreProps> = ({ className }) => {
     setDelegationPrimaryAddress: state.setDelegationPrimaryAddress,
     setDelegationSecondaryAddress: state.setDelegationSecondaryAddress,
     setIsDelegationCalculatingUQScoreModalOpen: state.setIsDelegationCalculatingUQScoreModalOpen,
-    setIsDelegationCompleted: state.setIsDelegationCompleted,
     setIsDelegationConnectModalOpen: state.setIsDelegationConnectModalOpen,
     setIsDelegationInProgress: state.setIsDelegationInProgress,
     setPrimaryAddressScore: state.setPrimaryAddressScore,
@@ -81,12 +72,7 @@ const HomeGridUQScore: FC<HomeGridUQScoreProps> = ({ className }) => {
       : primaryAddressScore === null || primaryAddressScore !== address,
   );
 
-  const { mutateAsync: checkDelegationMutation } = useCheckDelegation();
-  const {
-    mutateAsync: refreshAntisybilStatus,
-    isSuccess: isSuccessRefreshAntisybilStatus,
-    error: refreshAntisybilStatusError,
-  } = useRefreshAntisybilStatus();
+  const { error: refreshAntisybilStatusError } = useRefreshAntisybilStatus();
 
   const { data: antisybilStatusScore, isSuccess: isSuccessAntisybilStatusScore } =
     useAntisybilStatusScore(
@@ -94,10 +80,11 @@ const HomeGridUQScore: FC<HomeGridUQScoreProps> = ({ className }) => {
         ? delegationSecondaryAddress!
         : address!,
       {
-        enabled:
-          isSuccessRefreshAntisybilStatus ||
-          (refreshAntisybilStatusError as null | { message: string })?.message ===
-            'Address is already used for delegation',
+        /**
+         * Loading of the main antisybilStatus is done in useDelegationCheck,
+         * since it's required for onboarding too.
+         */
+        enabled: false,
       },
     );
 
@@ -124,45 +111,6 @@ const HomeGridUQScore: FC<HomeGridUQScoreProps> = ({ className }) => {
     };
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [isCalculatingYourUniquenessModalOpen]);
-
-  const checkDelegation = async () => {
-    if (!isUserTOSAccepted) {
-      return;
-    }
-    const accountsPromises = connectors.map(connector => connector.getAccounts());
-    /**
-     * In Safari browsers this array of promises results in an empty result,
-     * resulting in an error thrown from wagmi, ProviderNotFoundError.
-     *
-     * Adding empty catch here solves the problem.
-     */
-    const addresses = await Promise.all(accountsPromises).catch(() => {});
-    const uniqAddresses = addresses ? uniq(addresses.flat()) : [];
-    if (uniqAddresses.length < 2) {
-      refreshAntisybilStatus(address!);
-      return;
-    }
-    if (uniqAddresses.length > 10) {
-      refreshAntisybilStatus(address!);
-      toastService.showToast({
-        message: t('toasts.delegationTooManyUniqueAddresses.message'),
-        name: 'delegationTooManyUniqueAddresses',
-        title: t('toasts.delegationTooManyUniqueAddresses.title'),
-        type: 'error',
-      });
-      return;
-    }
-    checkDelegationMutation(uniqAddresses)
-      .then(({ primary, secondary }) => {
-        setDelegationPrimaryAddress(primary);
-        setDelegationSecondaryAddress(secondary);
-        setIsDelegationCompleted(true);
-        refreshAntisybilStatus(secondary);
-      })
-      .catch(() => {
-        refreshAntisybilStatus(address!);
-      });
-  };
 
   useEffect(() => {
     if (!isSuccessAntisybilStatusScore || isDelegationInProgress || isFetchingUqScore) {
@@ -206,19 +154,6 @@ const HomeGridUQScore: FC<HomeGridUQScoreProps> = ({ className }) => {
     };
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [isDelegationInProgress]);
-
-  useEffect(() => {
-    if (
-      (isDelegationCompleted &&
-        (delegationPrimaryAddress === address || delegationSecondaryAddress === address)) ||
-      !isUserTOSAccepted
-    ) {
-      return;
-    }
-    setIsDelegationCompleted(false);
-    checkDelegation();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [isUserTOSAccepted, address]);
 
   const isRecalculateButtonDisabled =
     !isConnected ||
